customer_service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic[email]
python-dotenv

